pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pareto-llm"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "psutil>=5.9",
    "huggingface-hub>=0.20",
]

[project.optional-dependencies]
mlx = [
    "mlx-lm>=0.20",
    "uvicorn>=0.20",
]
cuda = [
    "llama-cpp-python>=0.2",
    "nvidia-ml-py>=12.0",
    "uvicorn>=0.20",
    "starlette-context>=0.3",
    "fastapi>=0.100",
    "sse-starlette>=1.6",
    "pydantic-settings>=2.0",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]
terminal-bench = [
    "harbor>=0.1",
]
live-bench = [
    "livebench @ git+https://github.com/livebench/livebench.git@18b524d2d3a32282300b3548ef69e7d904ece836",
    "setuptools<71",  # livebench's syllapy dependency uses pkg_resources, removed in setuptools 71
    "en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl",  # livebench's instruction_following scorer
]

[project.scripts]
pareto-llm = "pareto_llm.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/pareto_llm"]

[tool.hatch.metadata]
allow-direct-references = true

[[tool.uv.index]]
name = "llama-cpp-python-cuda"
url = "https://abetlen.github.io/llama-cpp-python/whl/cu124"
explicit = true

[tool.uv.sources]
llama-cpp-python = [
    { index = "llama-cpp-python-cuda", extra = "cuda" },
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I"]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
]
